+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Initial i18n support [#20](https://github.com/opendatateam/udata-ods/pull/20)
 - Do not export SHP when above records limit [#26](https://github.com/opendatateam/udata-ods/pull/20)
 - Initial resources preview support [#29](https://github.com/opendatateam/udata-ods/pull/29) [#34](https://github.com/opendatateam/udata-ods/pull/34)
+- Use `udata.frontend.markdown:parse_html()` to parse description [#35](https://github.com/opendatateam/udata-ods/pull/35)
 
 ## 1.0.1 (2018-03-13)